% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntl_ice_doc.R
\docType{data}
\name{ntl_ice}
\alias{ntl_ice}
\title{North Temperate Lakes LTER: Ice Duration - Madison Lakes Area 1853 - current}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 431 rows and 6 columns.
}
\source{
{Magnuson, J.J., S.R. Carpenter, and E.H. Stanley. 2021. North Temperate Lakes LTER: Ice Duration - Madison Lakes Area 1853 - current ver 35. Environmental Data Initiative. https://doi.org/10.6073/pasta/ab31f2489ee436beb73fc8f1d0213d97 (Accessed 2021-03-01).}
\url{https://portal.edirepository.org/nis/mapbrowse?scope=knb-lter-ntl&identifier=33&revision=35}
}
\usage{
ntl_ice
}
\description{
Data includes lake name, dates of freeze-up and thaw, and duration of ice
cover of lakes in the Madison, WI area.
}
\keyword{datasets}
