% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwt_stress_doc.R
\docType{data}
\name{nwt_stress}
\alias{nwt_stress}
\title{Niwot Ridge LTER Pika Stress and Habitat Measurements}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 109 rows and 8 columns.
}
\source{

}
\usage{
nwt_stress
}
\description{
Niwot Ridge pika stress measurements collected from 2018-2019
}
\keyword{datasets}
