% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_abundance_doc.R
\docType{data}
\name{sbc_abundance}
\alias{sbc_abundance}
\title{Santa Barbara Coastal Fish Abundance in Kelp Removal Sites}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 74019 rows and 25 columns.
}
\source{
{Santa Barbara Coastal LTER and D. Reed. 2021. SBC LTER: Reef: Long-term experiment: Kelp removal: Fish abundance ver 20. Environmental Data Initiative. https://doi.org/10.6073/pasta/18408185dc13be279c20b2afcd8cb842 (Accessed 2021-02-13).}
\url{https://portal.edirepository.org/nis/mapbrowse?scope=knb-lter-sbc&identifier=30&revision=20}
}
\usage{
sbc_abundance
}
\description{
Number and Size of fish at the Santa Barbara Coastal LTER
}
\keyword{datasets}
