% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwt_pikas_doc.R
\docType{data}
\name{nwt_pikas}
\alias{nwt_pikas}
\title{Niwot Ridge LTER Pikas}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 376 rows and 33 columns.
}
\source{
{Ray, C. and Niwot Ridge LTER. 2021. Pika demography data for west knoll and Indian Peaks wilderness, 2008 - ongoing ver 4. Environmental Data Initiative. https://doi.org/10.6073/pasta/0a786c99fe3d4e1dfb8c57424ce79091 (Accessed 2021-02-15).}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-nwt.8.4}
}
\usage{
nwt_pikas
}
\description{
Niwot Ridge pika data collected during the summer months
}
\keyword{datasets}
